+++ src/app/guards/auth.guard.ts
@@ -1,7 +1,8 @@
-import { CanActivateFn, Router } from '@angular/router';
-import { AuthService } from '../services/auth.service';
 import { inject } from '@angular/core';
+import { CanActivateFn, Router } from '@angular/router';
 import { map } from 'rxjs';
+
+import { AuthService } from '../services/auth.service';
 
 export const authGuard: CanActivateFn = (route, state) => {
   const authService = inject(AuthService);