+++ src/app/interceptors/auth.interceptor.ts
@@ -1,8 +1,9 @@
-import { HttpInterceptorFn } from '@angular/common/http';
-import { AuthService } from '../services/auth.service';
 import { inject } from '@angular/core';
-import { catchError, switchMap, throwError } from 'rxjs';
+import { HttpInterceptorFn } from '@angular/common/http';
 import { Router } from '@angular/router';
+import { catchError, switchMap, throwError } from 'rxjs';
+
+import { AuthService } from '../services/auth.service';
 
 export const authInterceptor: HttpInterceptorFn = (req, next) => {
   const router = inject(Router);